working version of WL Migrations script 'from mooringlicensing.utils.waiting_list_migrate import WaitingListMigration; WaitingListMigration(test=False)'

+++ mooringlicensing/components/approvals/models.py
@@ -1859,11 +1859,12 @@ class StickerActionDetail(models.Model):
 
 @receiver(pre_delete, sender=Approval)
 def delete_documents(sender, instance, *args, **kwargs):
-    for document in instance.documents.all():
-        try:
-            document.delete()
-        except:
-            pass
+    if hasattr(instance, 'documents'):
+        for document in instance.documents.all():
+            try:
+                document.delete()
+            except:
+                pass
 
 
 #import reversion
